+++ src/main/java/com/team/buddyya/point/service/PointService.java
@@ -5,15 +5,11 @@ import com.team.buddyya.certification.domain.RegisteredPhone;
 import com.team.buddyya.certification.exception.PhoneAuthenticationException;
 import com.team.buddyya.certification.exception.PhoneAuthenticationExceptionType;
 import com.team.buddyya.certification.repository.RegisteredPhoneRepository;
-import com.team.buddyya.notification.service.NotificationService;
 import com.team.buddyya.point.domain.Point;
-import com.team.buddyya.point.domain.PointChangeType;
 import com.team.buddyya.point.domain.PointStatus;
 import com.team.buddyya.point.domain.PointType;
 import com.team.buddyya.point.dto.PointListResponse;
-import com.team.buddyya.point.dto.PointMissionResponse;
 import com.team.buddyya.point.dto.PointResponse;
-import com.team.buddyya.point.dto.PointMissionRewardResponse;
 import com.team.buddyya.student.domain.Student;
 import com.team.buddyya.point.repository.PointRepository;
 import com.team.buddyya.point.repository.PointStatusRepository;
@@ -31,12 +27,10 @@ import java.util.stream.Collectors;
 public class PointService {
 
     private final PointRepository pointRepository;
-    private final UpdatePointService updatePointService;
     private final PointStatusRepository pointStatusRepository;
     private final FindPointService findPointService;
     private final FindStudentService findStudentService;
     private final RegisteredPhoneRepository registeredPhoneRepository;
-    private final NotificationService notificationService;
 
     public Point createPoint(Student student) {
         RegisteredPhone registeredPhone = registeredPhoneRepository.findByPhoneNumber(student.getPhoneNumber())
@@ -72,33 +66,5 @@ public class PointService {
                 .map(PointResponse::from)
                 .collect(Collectors.toList());
         return PointListResponse.from(point, pointResponses);
-    }
-
-    public PointMissionResponse getUserMissionInfo(StudentInfo studentInfo){
-        Student student = findStudentService.findByStudentId(studentInfo.id());
-        RegisteredPhone registeredPhone = registeredPhoneRepository.findByPhoneNumber(student.getPhoneNumber())
-                .orElseThrow(() -> new PhoneAuthenticationException(PhoneAuthenticationExceptionType.PHONE_NOT_FOUND));
-        Point point = findPointService.findByStudent(student);
-        int totalMissionPoint = getTotalMissionPoint(point);
-        return PointMissionResponse.from(student, registeredPhone, totalMissionPoint);
-    }
-
-    public PointMissionRewardResponse checkAttendanceAndReward(StudentInfo studentInfo) {
-        Student student = findStudentService.findByStudentId(studentInfo.id());
-        RegisteredPhone registeredPhone = registeredPhoneRepository.findByPhoneNumber(student.getPhoneNumber())
-                .orElseThrow(() -> new PhoneAuthenticationException(PhoneAuthenticationExceptionType.PHONE_NOT_FOUND));
-        if (registeredPhone.isTodayAttended()) {
-            return PointMissionRewardResponse.from(null, PointType.NO_POINT_CHANGE.getPointChange());
-        }
-        Point point = updatePointService.updatePoint(student, PointType.MISSION_VISIT_REWARD);
-        registeredPhone.updateLastAttendanceDateToToday();
-        return PointMissionRewardResponse.from(point.getCurrentPoint(), PointType.MISSION_VISIT_REWARD.getPointChange());
-    }
-
-    public int getTotalMissionPoint(Point point) {
-        return pointStatusRepository.findAllByPointOrderByCreatedDateDesc(point).stream()
-                .filter(ps -> ps.getPointType().getChangeType() == PointChangeType.MISSION)
-                .mapToInt(PointStatus::getChangedPoint)
-                .sum();
     }
 }
